nvim: Alabasterize catppuccin highlighting, part 1
I'm starting to think that tonsky has a point in
https://tonsky.me/blog/syntax-highlighting/

Not sure I want to go _this_ minimal and I want to keep the non-syntax
goodness of a popular theme around. Catppuccin is easy to customize, so
I've copied the syntax highlighting from their repo, moved stuff around
and started to change highlights as I see fit.
Lua, ruby and c++ read... bearable at the moment. The rest will follow
(finding some weirdness in fugitive and git-commits as I'm typing this)